+++ desktop/cmp/grid/editors/impl/InlineEditorModel.js
@@ -11,6 +11,7 @@ import {HoistModel, useLocalModel} from '@xh/hoist/core';
 import {bindable, makeObservable} from '@xh/hoist/mobx';
 import {wait} from '@xh/hoist/promise';
 import {createObservableRef} from '@xh/hoist/utils/react';
+import {apiRemoved} from '@xh/hoist/utils/js';
 
 /**
  * Hook to render a component to be used for inline cell editing in ag-grid.
@@ -24,9 +25,11 @@ import {createObservableRef} from '@xh/hoist/utils/react';
  * @param {Object} ref - forwardRef passed to containing component
  * @return {ReactElement} - React Element to be rendered
  */
-export function useInlineEditorModel(component, props, ref) {
+export function useInlineEditorModel(component, props, ref, isPopup = false) {
     const {className, inputProps, agParams} = props,
         impl = useLocalModel(() => new InlineEditorModel(agParams));
+
+    apiRemoved('isPopup', {test: isPopup, msg: 'Use `Column.editorIsPopup` instead', v: 48});
 
     useImperativeHandle(ref, () => ({
         getValue: () => impl.value,